Added globalVolume to the Channel Contents grammar rule since that command can occur in a channel Added more messages for errors and warnings Fixed the CheckBits to use the correct values Added additional logic to process AMK and Pan commands Adjust the CatalogueUserDefinedInformation function to track named and unnamed non-superloops Fixed logic pertaining to various Loop Evaluations Updated various minor pieces of logic, messages, etc

+++ Addmusic2/Model/ChannelInformation.cs
@@ -11,7 +11,8 @@ namespace Addmusic2.Model
 
         public int ChannelNumber { get; set; }
         public List<byte> ChannelData { get; set; } = new();
-        public List<byte> LoopLocations { get; set; } = new();
+        //public List<byte> LoopLocations { get; set; } = new();
+        public List<ushort> LoopLocations { get; set; } = new();
         public double ChannelLength { get; set; }
         public bool HasIntro { get; set; } = false;
         public byte PhraseLocation { get; set; } = 0;
